# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Barbearia",
        "description": "Documentação dos endpoints da barbearia desenvolvida em Laravel 11",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local (Docker)"
        }
    ],
    "paths": {
...
        "/barbeiros": {
            "get": {
                "tags": [
                    "Barbeiros"
                ],
                "summary": "Listar todos os barbeiros",
                "description": "Método responsável por listar todos os barbeiros cadastrados no estabelecimento.",
                "operationId": "ac18858114beec5e4e22c14826d5b86d",
                "responses": {
                    "200": {
                        "description": "Lista de barbeiros retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Barbeiros"
                ],
                "summary": "Cadastrar novo barbeiro",
                "description": "Método responsável por cadastrar e persistir um novo barbeiro/profissional no sistema.",
                "operationId": "d795f590f55526dd97525966db333f33",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "especialidade"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Navalha de Ouro"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Barbeiro cadastrado com sucesso"
                    }
                }
            }
        },
        "/barbeiros/{id}": {
            "get": {
                "tags": [
                    "Barbeiros"
                ],
                "summary": "Obter dados de um barbeiro específico",
                "description": "Método responsável por buscar e exibir os dados de um barbeiro específico por meio de seu ID.",
                "operationId": "0444948625b5b1ee2741751418b78a03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID numérico do barbeiro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do barbeiro retornados com sucesso"
                    },
                    "404": {
                        "description": "Barbeiro não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Barbeiros"
                ],
                "summary": "Excluir um barbeiro",
                "description": "Método responsável por remover permanentemente um barbeiro do banco de dados do sistema.",
                "operationId": "ab3355f891e150faaf288243c6223626",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID numérico do barbeiro a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barbeiro removido com sucesso"
                    },
                    "404": {
                        "description": "Barbeiro não encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Barbeiros"
                ],
                "summary": "Atualizar dados de um barbeiro",
                "description": "Método responsável por atualizar a especialidade ou outros dados cadastrais de um barbeiro existente.",
                "operationId": "10f7a1f8b2286be3931cf9af1055bc85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID numérico do barbeiro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "especialidade": {
                                        "type": "string",
                                        "example": "Especialista em Degradê e Progressiva"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barbeiro atualizado com sucesso"
                    },
                    "404": {
                        "description": "Barbeiro não encontrado"
                    }
                }
            }
        },
        "/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos os clientes",
                "description": "Método responsável por listar todos os clientes cadastrados na base de dados.",
                "operationId": "7e6daa583638d8c1239ebb00020b6b0d",
                "responses": {
                    "200": {
                        "description": "Lista de clientes retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastrar novo cliente",
                "description": "Método responsável por criar e persistir um novo cliente no sistema.",
                "operationId": "42d1ed0a3b662bee8adb649a163ac272",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "telefone",
                                    "email"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Alisson Sommer"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "47999998888"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "alisson@email.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente cadastrado com sucesso"
                    }
                }
            }
        },
        "/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obter dados de um cliente específico",
                "description": "Método responsável por buscar e exibir os detalhes de um cliente específico através do seu código identificador.",
                "operationId": "37e8a9756593ca4a25faafb81dba175c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID numérico do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do cliente retornados com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Excluir um cliente do sistema",
                "description": "Método responsável por remover definitivamente um cliente do banco de dados e retornar a confirmação do código excluído.",
                "operationId": "04f02157546cfad8ff244a1be6768d8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID numérico do cliente a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente excluído com sucesso e ID retornado"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualizar dados de um cliente",
                "description": "Método responsável por atualizar de forma parcial ou total os dados cadastrais de um cliente existente.",
                "operationId": "8364a21e48efa5a85222dac16006ced8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID numérico do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Alisson S. Mudado"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "47988887777"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            }
        },
...
    },
...
}